% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_stats.R
\name{add_k10_change}
\alias{add_k10_change}
\title{Calculates reliable change for k10 scores}
\usage{
add_k10_change(data, pre, post, pop_sd = 9.2, rel = 0.92)
}
\arguments{
\item{data}{a dataframe in wide format}

\item{pre}{the column containing the pre-measure}

\item{post}{the column containing a post measure}

\item{pop_sd}{a revised estimate for the population standard deviation}

\item{rel}{a revised estimate from the measure reliability}
}
\value{
the input dataframe with a raw change and a reliable change indicator appended
}
\description{
Simple differences in pre- and post- scores risks overinterpreting variation.
Variation in outcome measures are to be expected due to uncontrolled factors
unrelated to a consumers' actual outcomes. Reliable change indices
/url(https://ebmh.bmj.com/content/1/3/70) seek to place benchmarks on changes
that are attributable to random noise and those that are more systematic.
This calculation relies on an estimate of the standard deviation (AMHOCN
estimate 9.2, n =281; Neami intake estimate 8.84, n = 2215) and of scale
reliability (alpha = .91: https://rdcu.be/cW31L, neami estimate alpha = .92)
}
\examples{
data <-  data.frame(person = 1:30, pre = rnorm(30, 30, 9.2), post = rnorm(30, 22, 9.2))
add_k10_change(data, pre, post)
}
