% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_stats.R
\name{add_k5_change}
\alias{add_k5_change}
\title{Calculates reliable change for k5 scores}
\usage{
add_k5_change(data, pre, post)
}
\arguments{
\item{data}{a dataframe in wide format}

\item{pre}{the column containing the pre-measure}

\item{post}{the column containing a post measure}
}
\value{
the input dataframe with a raw change and a reliable change indicator appended
}
\description{
Simple differences in pre- and post- scores risks overinterpreting variation.
Variation in outcome measures are to be expected due to uncontrolled factors
unrelated to a consumers' actual outcomes. Reliable change indices
/url(https://ebmh.bmj.com/content/1/3/70) seek to place benchmarks on changes
that are attributable to random noise and those that are more systematic.
The critical value in this function was provided by the department of health 2024-25
}
\examples{
data <-  data.frame(person = 1:30, pre = rnorm(30, 15, 7), post = rnorm(30, 11, 7))
add_k5_change(data, pre, post)
}
