% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_stats.R
\name{add_rasds_change}
\alias{add_rasds_change}
\title{Calculates reliable change for ras-ds scores}
\usage{
add_rasds_change(data, pre, post, pop_sd = 20, rel = 0.94)
}
\arguments{
\item{data}{a dataframe in wide format}

\item{pre}{the column containing the pre-measure}

\item{post}{the column containing a post measure}

\item{pop_sd}{a revised estimate for the population standard deviation}

\item{rel}{a revised estimate from the measure reliability}
}
\value{
the input dataframe with a raw change and a reliable change indicator appended
}
\description{
Simple differences in pre- and post- scores risks overinterpreting variation.
Variation in outcome measures are to be expected due to uncontrolled factors
unrelated to a consumers' actual outcomes. Reliable change indices
/url(https://ebmh.bmj.com/content/1/3/70) seek to place benchmarks on changes
that are attributable to random noise and those that are more systematic.
This calculation relies on an estimate of the standard deviation (Neami 
intake estimate 20, n ~= 900) and of scale  reliability (neami estimate alpha = .94)
}
\examples{
data <-  data.frame(person = 1:30, pre = rnorm(30, 85, 20), post = rnorm(30, 115, 20))
add_rasds_change(data, pre, post)
}
