% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_age.R
\name{bin_age_10}
\alias{bin_age_10}
\title{Bins ages against ABS age standard}
\usage{
bin_age_10(age)
}
\arguments{
\item{age}{a vector of integer ages}
}
\value{
a factor vector of 10 year age brackets
}
\description{
Bins ages against ABS age standard
}
\examples{
bin_age_10(sample(2:90, 200))
}
