% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_help.R
\name{bounding_box}
\alias{bounding_box}
\title{Calculate Bounding Box}
\usage{
bounding_box(lat, lon, dist, in.km = TRUE)
}
\arguments{
\item{lat}{The latitude of the center point.}

\item{lon}{The longitude of the center point.}

\item{dist}{The distance from the center point.}

\item{in.km}{logical.  If \code{TRUE} uses km as the units of
\code{dist}.  If \code{FALSE} uses miles.}
}
\value{
Returns a matrix with max/min latitude/longitude values.
}
\description{
Calculate a bounding box for a center point given a set of coordinates.
}
\examples{
bounding_box(-37.74348546926522, 145.03134441138408, 1)
}
\references{
\url{http://janmatuschek.de/LatitudeLongitudeBoundingCoordinates}
}
\keyword{bounding_box,}
\keyword{coordinates}
