% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{caption_ttest}
\alias{caption_ttest}
\title{Wraps the output of a t-test into a pretty format}
\usage{
caption_ttest(ttest)
}
\arguments{
\item{ttest}{An S3 t-test model object}
}
\value{
A character string of the model summary
}
\description{
Wraps the output of a t-test into a pretty format
}
