% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coders.R
\name{code_cald}
\alias{code_cald}
\title{Codes Persons into CALD status}
\usage{
code_cald(country_of_birth, main_language)
}
\arguments{
\item{country_of_birth}{a vector of countries of birth}

\item{main_language}{a vector of main language spoken at home}
}
\value{
a vector of cald statuses
}
\description{
Defininc CALD communities is difficult. Per Pham et al. (2021), 
for routine data projects, Neami presumes consumers come from a 
Culturally or Linguistically Diverse Community if their Main 
language is not English and their country of Birth is not an 
English majority-speaking country. This will underestimate 
cultural diversity.
}
\examples{
 code_cald(country_of_birth =c("England", "England", "Narnia", "Spain", NA_character_), 
 main_language = c("English", "Spanish", "French", "English", NA_character_))
}
