% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coders.R
\name{code_contacts}
\alias{code_contacts}
\title{Code PMHC service contact types against reference codes}
\usage{
code_contacts(guid)
}
\arguments{
\item{guid}{a vector of service contact contact IDs}
}
\value{
a vector of contact types
}
\description{
Code PMHC service contact types against reference codes
}
\examples{
code_contacts(c("33A9B8BA-341C-4D25-B5C0-2A7DAE8C8AB6", "73F89E5F-AF10-47DD-B3DC-37018D1DA6A7"))
}
