% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_NPN.R
\name{config_vpn}
\alias{config_vpn}
\title{Configure VPN settings}
\usage{
config_vpn(vpn = NULL, password = NULL)
}
\arguments{
\item{vpn}{The name of the VPN}

\item{password}{User Password}
}
\description{
Walks a user through setting VPN parameters to use the `login_vpn()` function to connect to the Neami VPN. Credentials are held on system.
}
