% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_note_type.R
\name{decode_note_type}
\alias{decode_note_type}
\title{Convert carelink note type codes to plaintext}
\usage{
decode_note_type(code)
}
\arguments{
\item{code}{a character vector containing a note type code}
}
\value{
a character vector of plaintext note types.
}
\description{
Convert carelink note type codes to plaintext
}
\examples{
decode_note_type("F58199AF-5E52-4514-93D4-DF86D963BF93")
}
