% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojify.R
\name{emojify}
\alias{emojify}
\title{Converts content to emoji}
\usage{
emojify(x, type = "TF")
}
\arguments{
\item{x}{a logical vector}

\item{type}{a parameter that will allow for alternative codings.}
}
\value{
a vector of emoji
}
\description{
Converts content to emoji
}
\examples{
emojify(c(TRUE, FALSE, TRUE))
}
