% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{finalise_plot}
\alias{finalise_plot}
\title{Arrange alignment and save Neami ggplot chart}
\usage{
finalise_plot(
  plot_name,
  source_name,
  save = TRUE,
  save_filepath = file.path(getwd(), "tmp-nc.png"),
  width_pixels = 640,
  height_pixels = 450,
  logo_image_path = system.file("logos/NN.master.Large_nostrap.png", package = "neami")
)
}
\arguments{
\item{plot_name}{The variable name of the plot you have created that you want to format and save}

\item{source_name}{The text you want to come after the text 'Source:' in the bottom left hand side of your side}

\item{save}{A logical which indicates if the plot should be saved}

\item{save_filepath}{Exact filepath that you want the plot to be saved to}

\item{width_pixels}{Width in pixels that you want to save your chart to - defaults to 640}

\item{height_pixels}{Height in pixels that you want to save your chart to - defaults to 450}

\item{logo_image_path}{File path for the logo image you want to use in the right hand side of your chart,
which needs to be a PNG file - defaults to Neami Logo image that sits within the logos folder of your package}
}
\value{
(Invisibly) an updated ggplot object.
}
\description{
Running this function will save your plot with the correct guidelines for publication by Neami.
It will left align your title, subtitle and source, add the Neami Logo at the bottom right and save it to your specified location.
}
\examples{
finalise_plot(plot_name = myplot,
source = "The source for my data",
save_filepath = "filename_that_my_plot_should_be_saved_to-nc.png",
width_pixels = 640,
height_pixels = 450,
logo_image_path = "logo_image_filepath.png"
)

}
\keyword{finalise_plot}
