% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_stats.R
\name{flag_k10_changes}
\alias{flag_k10_changes}
\title{Flags reliable changes to k10s in a long data frame}
\usage{
flag_k10_changes(data, name = name, value = value, levels = c("pre", "post"))
}
\arguments{
\item{data}{a long data frame containing 2 observations per consumer}

\item{name}{the unquoted column name containing timepoint names}

\item{value}{the unquoted column name containing timepoint values}

\item{levels}{a vector of timepoint labels ordered ascending}
}
\description{
Flags reliable changes to k10s in a long data frame
}
\examples{
data <-  data.frame(person = 1:30, pre = rnorm(30, 30, 9.2), post = rnorm(30, 22, 9.2))
dl <- tidyr::pivot_longer(data,  pre:post)
flag_k10_changes(dl, name, value, levels = c("pre", "post"))
}
