% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_tools.R
\name{inspire_coder}
\alias{inspire_coder}
\title{Codes Inspire O items}
\usage{
inspire_coder(item)
}
\arguments{
\item{item}{a vector of Inspire O response character strings in Sentence case}
}
\value{
a vector of integer values
}
\description{
Codes Inspire O items
}
