% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_tools.R
\name{k10_coder}
\alias{k10_coder}
\title{Codes K10/k5 items}
\usage{
k10_coder(item)
}
\arguments{
\item{item}{a vector of k10/k5 response character strings in Sentence case}
}
\value{
a vector of integer values
}
\description{
Codes K10/k5 items
}
