% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_episodes.R
\name{list_periods}
\alias{list_periods}
\title{List date units between a start and end date}
\usage{
list_periods(from, to, by = "month")
}
\arguments{
\item{from}{the start date (will be floored to the `by` unit)}

\item{to}{the end date}

\item{by}{the units the span should be counted by (e.g. "month", "quarter")}
}
\value{
A vector of dates
}
\description{
List date units between a start and end date
}
\examples{
list_periods(ymd("2024-06-01"), Sys.Date())
}
