% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_NPN.R
\name{login_vpn}
\alias{login_vpn}
\title{Login to Neami VPN}
\usage{
login_vpn()
}
\description{
Provides a function to establish a connection with a VPN. Requires configuration using the `config_vpn()` function.
}
\examples{
login_vpn()
}
