% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_rename.R
\name{meta_rename}
\alias{meta_rename}
\title{Replaces names by reference to a metadata dataframe}
\usage{
meta_rename(df, metadata, old, new)
}
\arguments{
\item{df}{A data.frame object for renaming.}

\item{metadata}{A data.frame object containing the key-value pair}

\item{old}{An unquoted string of the key column name in the metadata object}

\item{new}{An unquoted string of the value column name in the metadata object}
}
\value{
`data.frame` object with new names as per key-value pairs.
}
\description{
`meta_rename` allows fast replacement of variable names by key-value pairing with a metadata file.
}
