% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmex_helpers.R
\name{mmex_birth_country_code}
\alias{mmex_birth_country_code}
\title{Recodes country of birth from PMHC codes to names}
\usage{
mmex_birth_country_code(x)
}
\arguments{
\item{x}{a vector of integers indexing countries of birth}
}
\value{
a vector of names
}
\description{
Recodes country of birth from PMHC codes to names
}
