% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmex_helpers.R
\name{mmex_eng_prof}
\alias{mmex_eng_prof}
\title{Recodes English proficiency}
\usage{
mmex_eng_prof(x)
}
\arguments{
\item{x}{a vector of integers indexing proficiency in spoken english}
}
\value{
a vector of values
}
\description{
Recodes English proficiency
}
