% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmex_helpers.R
\name{mmex_home_lang}
\alias{mmex_home_lang}
\title{Recodes main language at home from PMHC codes to names}
\usage{
mmex_home_lang(x)
}
\arguments{
\item{x}{a vector of integers indexing main language spoken at home}
}
\value{
a vector of names
}
\description{
Recodes main language at home from PMHC codes to names
}
