% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{neami_pal}
\alias{neami_pal}
\title{neami palette with ramped colours}
\usage{
neami_pal(palette = "neami_qual", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'neami_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
neami palette with ramped colours
}
\examples{
library(scales)
show_col(ochre_pal()(10))

filled.contour(volcano,color.palette = neami_pal(), asp=1)

}
