% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{neami_palettes}
\alias{neami_palettes}
\title{Neami National Colour Palettes}
\format{
An object of class \code{list} of length 4.
}
\usage{
neami_palettes
}
\description{
A collection of Neami colour palettes. The list of available palettes is:
neami_qual
neami_seq1
neami_seq2
neami_diverging
}
\examples{

# Make an x-y plot using the neami_diverging palette
library(tidyverse)
df <- data.frame(x = rnorm(100, 0, 20),
          y = rnorm(100, 0, 20),
          cl = sample(letters[1:8], 100, replace=TRUE))
ggplot(df, aes(x, y, colour=cl, shape=cl)) +
  geom_point(size=4) + scale_colour_neami() +
  theme_bw() + theme(aspect.ratio=1)
}
\keyword{datasets}
