% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_tools.R
\name{pair_outcome_measures}
\alias{pair_outcome_measures}
\title{Filters and pairs outcome measures}
\usage{
pair_outcome_measures(
  df,
  grouping_id,
  measure_total,
  collection_occasions,
  collection_date,
  min_interval = 30
)
}
\arguments{
\item{df}{a dataframe}

\item{grouping_id}{a column name indicating the episode key / person id}

\item{measure_total}{a column name indicating the measure}

\item{collection_occasions}{a column name indicating the collection occasion}

\item{collection_date}{a column name indicating the collation date}

\item{min_interval}{a integer of days between measures for valid pairing}
}
\value{
a data frame of paired measures more than an arbirtary number of days apart
}
\description{
Filters and pairs outcome measures
}
