% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{paired_t_plot}
\alias{paired_t_plot}
\title{Create a plot to represent a paired t-test}
\usage{
paired_t_plot(data, x, y, group)
}
\arguments{
\item{data}{a dataframe in long format}

\item{x}{a bare variable name containing the timepoint indicator}

\item{y}{a bare variable name containing the values for comparison}

\item{group}{a bare variable name containing the case ID}
}
\description{
Create a plot to represent a paired t-test
}
