% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{paired_t_test}
\alias{paired_t_test}
\title{Implements a paired t-test from a long dataframe}
\usage{
paired_t_test(df, outcome, timeid)
}
\arguments{
\item{df}{a long dataframe or tibble}

\item{outcome}{the continuous numerical outcome variable for comparison}

\item{timeid}{the variable indexing the timepoint}
}
\description{
Implements a paired t-test from a long dataframe
}
