% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_tools.R
\name{prep_measures}
\alias{prep_measures}
\title{Prepares raw long measures from query into usable wide data frames}
\usage{
prep_measures(measures, fundings, type)
}
\arguments{
\item{measures}{a long dataframe containing one measure}

\item{fundings}{a data frame containing fundings}

\item{type}{a character indicating the measure type}
}
\value{
a wide data frame of forms linked to fundings
}
\description{
Prepares raw long measures from query into usable wide data frames
}
