% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_tools.R
\name{ras_coder}
\alias{ras_coder}
\title{Codes RAS-DS items}
\usage{
ras_coder(item)
}
\arguments{
\item{item}{a vector of ras-ds response character strings in Sentence case}
}
\value{
a vector of integer values
}
\description{
Codes RAS-DS items
}
