% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{save_plot}
\alias{save_plot}
\title{Saves a neami themed plot}
\usage{
save_plot(plot_grid, width, height, save_filepath)
}
\arguments{
\item{plot_grid}{The ggplot object for saving}

\item{width}{the output width in default units}

\item{height}{the output height in default units}

\item{save_filepath}{path to save output file to}
}
\description{
Saves a neami themed plot
}
