% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_colour_neami_c}
\alias{scale_colour_neami_c}
\title{Colour a geom with a continuous Neami palette}
\usage{
scale_colour_neami_c(name)
}
\arguments{
\item{name}{a character name of a palette}
}
\description{
Colour a geom with a continuous Neami palette
}
