% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_colour_neami_d}
\alias{scale_colour_neami_d}
\title{Colour a geom with a discrete Neami palette}
\usage{
scale_colour_neami_d(name)
}
\arguments{
\item{name}{a character name of a palette}
}
\description{
Colour a geom with a discrete Neami palette
}
