% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_fill_neami_c}
\alias{scale_fill_neami_c}
\title{Fill a geom with a continuous Neami palette}
\usage{
scale_fill_neami_c(name)
}
\arguments{
\item{name}{a character name of a palette}
}
\description{
Fill a geom with a continuous Neami palette
}
