% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_tools.R
\name{sdq_coder}
\alias{sdq_coder}
\title{Codes sdq items}
\usage{
sdq_coder(question, type, reverse = FALSE)
}
\arguments{
\item{question}{a vector of SDQ response character strings}

\item{type}{a scalar indicating the type of question "tf" for true/false, "impact" for impact}

\item{reverse}{a Logical indicates the item should be reverse coded}
}
\value{
a numeric vector of coded responses
}
\description{
Codes sdq items
}
