% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_help.R
\name{site_map}
\alias{site_map}
\title{Generate a map of consumers origin around a site}
\usage{
site_map(lat, lon, dist = 15, suburb_df, suburb, shapefile_path, state)
}
\arguments{
\item{lat}{Numeric Latitude of the site}

\item{lon}{Numeric Longitude of the site}

\item{dist}{Numeric Distance in KM around the site that the bounding box should be drawn}

\item{suburb_df}{Dataframe containing consumers and a column of suburbs}

\item{suburb}{The column of suburbs within the dataframe containing suburn names}

\item{shapefile_path}{Path to the shapefile to join.}

\item{state}{State of Australia to filter the shapefile to.}
}
\description{
Generate a map of consumers origin around a site
}
