% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_tools.R
\name{standardise_measures}
\alias{standardise_measures}
\title{Standardises measure occasion representations}
\usage{
standardise_measures(item, type = "occasion")
}
\arguments{
\item{item}{a character vector indicating the parameter for standardisation}

\item{type}{a string indicating the type of variable for coding}
}
\value{
a standardised vector of strings.
}
\description{
Standardises measure occasion representations
}
