% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_neami}
\alias{theme_neami}
\title{Neami ggplot2 theme}
\usage{
theme_neami()
}
\description{
The function theme_neami will apply a coherent "Neami" theme to a ggplot2 object.
}
\examples{
ggplot2::ggplot(mtcars, aes())
}
