% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_neami_dark}
\alias{theme_neami_dark}
\title{Neami ggplot2 dark theme}
\usage{
theme_neami_dark()
}
\description{
The function theme_neami_dark will apply a coherent "Neami" dark theme to a ggplot2 object.
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes()) + theme_neami_dark()
}
