% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_neami_map}
\alias{theme_neami_map}
\title{Neami ggplot2 theme for maps}
\usage{
theme_neami_map()
}
\description{
Applies the Neami theme to a ggplot2 map.
}
\examples{
library(maps)
world1 <- sf::st_as_sf(map('world', plot = FALSE, fill = TRUE))
ggplot() + geom_sf(data = world1) + theme_neami_map()
}
