% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{left_align}
\alias{left_align}
\title{Left align plot elements for themed plot}
\usage{
left_align(plot_name, pieces)
}
\arguments{
\item{plot_name}{the plot object for alignment}

\item{pieces}{a vector of parameter names for alignment e.g. c("subtitle", "title", "caption")}
}
\description{
Left align plot elements for themed plot
}
