% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{neami_colours}
\alias{neami_colours}
\title{Neami National Colour Palettes}
\format{
An object of class \code{list} of length 7.
}
\usage{
neami_colours
}
\description{
A collection of Neami colour palettes. The list of available palettes is:
core_qual_light
core_qual
blues
greens_l
oranges
}
\examples{

# Make an x-y plot using the neami_diverging palette
library(tidyverse)
df <- data.frame(x = rnorm(100, 0, 20),
          y = rnorm(100, 0, 20),
          cl = sample(letters[1:8], 100, replace=TRUE))
ggplot(df, aes(x, y, colour=cl)) +
  geom_point(size=4) + scale_colour_neami_d() +
  theme_bw() + theme(aspect.ratio=1)
}
\keyword{datasets}
