% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{neami_palettes}
\alias{neami_palettes}
\title{Retrieve a Neami palette}
\usage{
neami_palettes(
  name,
  n,
  all_palettes = neami_colours,
  type = c("discrete", "continuous")
)
}
\arguments{
\item{name}{name of palette}

\item{n}{number of levels}

\item{all_palettes}{list of named palettes}

\item{type}{a character entry indicating whether the palette is discrete or continuous}
}
\description{
Retrieve a Neami palette
}
