% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{neami_waffle}
\alias{neami_waffle}
\title{Makes an opinionated waffleplot}
\usage{
neami_waffle(data, count_var, palette = "Set2")
}
\arguments{
\item{data}{a dataframe}

\item{count_var}{a bare variable name indicating the column to be counted}

\item{palette}{a ggplot2 compatable pallette}
}
\value{
a waffleplot
}
\description{
Makes an opinionated waffleplot
}
\examples{
neami_waffle(mtcars, cyl)
}
